// backend/src/modules/vendors/vendor.types.ts
export enum VendorCategory {
  CIVIL = 'CIVIL',
  ARCHITECT = 'ARCHITECT',
  COTTAGE_MANUFACTURER = 'COTTAGE_MANUFACTURER',
  LANDSCAPING = 'LANDSCAPING',
  POOL = 'POOL',
  ELECTRICAL = 'ELECTRICAL',
  PLUMBING = 'PLUMBING',
  FURNITURE = 'FURNITURE',
  SUPPLIER = 'SUPPLIER',
  MARKETING = 'MARKETING',
  OTHER = 'OTHER',
}

export enum VendorWorkStatus {
  NOT_STARTED = 'NOT_STARTED',
  IN_PROGRESS = 'IN_PROGRESS',
  ON_HOLD = 'ON_HOLD',
  COMPLETED = 'COMPLETED',
  CANCELLED = 'CANCELLED',
}

export enum VendorPaymentStatus {
  NOT_PAID = 'NOT_PAID',
  PARTIALLY_PAID = 'PARTIALLY_PAID',
  PAID = 'PAID',
}